pyproject.toml
[project]
name = "dspy-hacking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=3.0.3",
    "python-dotenv>=1.1.1",
]
